Show job name in job spec details pages
