+++ operator_ui/src/pages/Jobs/RegionalNav.tsx
@@ -24,7 +24,7 @@ import Link from 'components/Link'
 import ErrorMessage from 'components/Notifications/DefaultError'
 import jobSpecDefinition from 'utils/jobSpecDefinition'
 import { isWebInitiator } from 'utils/jobSpecInitiators'
-import { JobData } from './Show'
+import { JobData } from './sharedTypes'
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -215,11 +215,6 @@ const RegionalNavComponent = ({
       <Card className={classes.container}>
         <Grid container spacing={0}>
           <Grid item xs={12}>
-            <Typography variant="subtitle2" color="secondary" gutterBottom>
-              Job Spec Detail
-            </Typography>
-          </Grid>
-          <Grid item xs={12}>
             <Grid
               container
               spacing={0}
@@ -232,7 +227,7 @@ const RegionalNavComponent = ({
                   color="secondary"
                   className={classes.jobSpecId}
                 >
-                  {jobSpecId}
+                  {(job && job.attributes.name) || jobSpecId}
                 </Typography>
               </Grid>
               <Grid item xs={6} className={classes.actions}>
@@ -272,6 +267,11 @@ const RegionalNavComponent = ({
             </Grid>
           </Grid>
           <Grid item xs={12}>
+            {job && job.attributes.name && (
+              <Typography variant="subtitle2" color="secondary" gutterBottom>
+                {job.attributes.id}
+              </Typography>
+            )}
             <Typography variant="subtitle2" color="textSecondary">
               Created{' '}
               {job && job.attributes.createdAt && (
